upgrade to tiptap v3 (#226)
* update to v3

* add tippyjs

* remove commented code

* fix full width of the menu items

* fix bubble menu rendering top-left after each transaction

The 'modal' prop was removed from the Popover component in EditorSelector, likely to adjust its behavior or appearance. This change may affect how the popover interacts with the rest of the UI.

* Remove shouldRerenderOnTransaction prop from EditorProvider

Eliminates the shouldRerenderOnTransaction prop from the EditorProvider component, likely to prevent unnecessary re-renders or because it is no longer needed.

* Fix formatting

---------

Co-authored-by: Hayden Bleasel <hello@haydenbleasel.com>